Core/Entities: First batch of removing implicit conversions of ObjectGuid to uint64

+++ src/server/game/Chat/Channels/Channel.cpp
@@ -552,7 +552,7 @@ void Channel::List(Player const* player)
              member->GetSession()->GetSecurity() <= AccountTypes(gmLevelInWhoList)) &&
             member->IsVisibleGloballyFor(player))
         {
-            data << uint64(i->first);
+            data << i->first;
             data << uint8(i->second.flags);             // flags seems to be changed...
             ++count;
         }
@@ -762,13 +762,13 @@ void Channel::MakeNotifyPacket(WorldPacket* data, uint8 notify_type)
 void Channel::MakeJoined(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_JOINED_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeLeft(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_LEFT_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeYouJoined(WorldPacket* data)
@@ -804,13 +804,13 @@ void Channel::MakeNotModerator(WorldPacket* data)
 void Channel::MakePasswordChanged(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_PASSWORD_CHANGED_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeOwnerChanged(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_OWNER_CHANGED_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakePlayerNotFound(WorldPacket* data, std::string const& name)
@@ -838,7 +838,7 @@ void Channel::MakeChannelOwner(WorldPacket* data)
 void Channel::MakeModeChange(WorldPacket* data, ObjectGuid guid, uint8 oldflags)
 {
     MakeNotifyPacket(data, CHAT_MODE_CHANGE_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
     *data << uint8(oldflags);
     *data << uint8(GetPlayerFlags(guid));
 }
@@ -846,13 +846,13 @@ void Channel::MakeModeChange(WorldPacket* data, ObjectGuid guid, uint8 oldflags)
 void Channel::MakeAnnouncementsOn(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_ANNOUNCEMENTS_ON_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeAnnouncementsOff(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_ANNOUNCEMENTS_OFF_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeMuted(WorldPacket* data)
@@ -863,8 +863,8 @@ void Channel::MakeMuted(WorldPacket* data)
 void Channel::MakePlayerKicked(WorldPacket* data, ObjectGuid bad, ObjectGuid good)
 {
     MakeNotifyPacket(data, CHAT_PLAYER_KICKED_NOTICE);
-    *data << uint64(bad);
-    *data << uint64(good);
+    *data << bad;
+    *data << good;
 }
 
 void Channel::MakeBanned(WorldPacket* data)
@@ -875,15 +875,15 @@ void Channel::MakeBanned(WorldPacket* data)
 void Channel::MakePlayerBanned(WorldPacket* data, ObjectGuid bad, ObjectGuid good)
 {
     MakeNotifyPacket(data, CHAT_PLAYER_BANNED_NOTICE);
-    *data << uint64(bad);
-    *data << uint64(good);
+    *data << bad;
+    *data << good;
 }
 
 void Channel::MakePlayerUnbanned(WorldPacket* data, ObjectGuid bad, ObjectGuid good)
 {
     MakeNotifyPacket(data, CHAT_PLAYER_UNBANNED_NOTICE);
-    *data << uint64(bad);
-    *data << uint64(good);
+    *data << bad;
+    *data << good;
 }
 
 void Channel::MakePlayerNotBanned(WorldPacket* data, const std::string &name)
@@ -895,13 +895,13 @@ void Channel::MakePlayerNotBanned(WorldPacket* data, const std::string &name)
 void Channel::MakePlayerAlreadyMember(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_PLAYER_ALREADY_MEMBER_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeInvite(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_INVITE_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeInviteWrongFaction(WorldPacket* data)
@@ -954,19 +954,19 @@ void Channel::MakeNotInLfg(WorldPacket* data)
 void Channel::MakeVoiceOn(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_VOICE_ON_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::MakeVoiceOff(WorldPacket* data, ObjectGuid guid)
 {
     MakeNotifyPacket(data, CHAT_VOICE_OFF_NOTICE);
-    *data << uint64(guid);
+    *data << guid;
 }
 
 void Channel::JoinNotify(ObjectGuid guid)
 {
     WorldPacket data(IsConstant() ? SMSG_USERLIST_ADD : SMSG_USERLIST_UPDATE, 8 + 1 + 1 + 4 + GetName().size());
-    data << uint64(guid);
+    data << guid;
     data << uint8(GetPlayerFlags(guid));
     data << uint8(GetFlags());
     data << uint32(GetNumPlayers());
@@ -981,7 +981,7 @@ void Channel::JoinNotify(ObjectGuid guid)
 void Channel::LeaveNotify(ObjectGuid guid)
 {
     WorldPacket data(SMSG_USERLIST_REMOVE, 8 + 1 + 4 + GetName().size());
-    data << uint64(guid);
+    data << guid;
     data << uint8(GetFlags());
     data << uint32(GetNumPlayers());
     data << GetName();
